Consolidate resources, remove interface files
