--- src/angular-resource-analyzer/angular-resource-url.interface.ts
@@ -1,10 +1,0 @@
-import * as typescript from 'typescript';
-
-/**
- * Angular resource URL interface
- */
-export interface AngularResourceUrl {
-	url: string;
-	node: typescript.Node;
-	content?: string; // Not part of the initial analysis result, added later on
-}
