--- src/angular-resource-analyzer/angular-resource.interface.ts
@@ -1,13 +1,0 @@
-import * as typescript from 'typescript';
-
-import { AngularResourceUrl } from './angular-resource-url.interface';
-
-/**
- * Angular Resource Interface
- */
-export interface AngularResource {
-	oldKey: string;
-	newKey: string;
-	node: typescript.Node;
-	urls: Array<AngularResourceUrl>;
-}
